requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
